test(s3stream): add MemoryObjectStorage for test
